Add data structure for objects state save
At the end for the replay system we are going a use a binary file, this will give better performance for storing and reading the information. We will have to be careful with the what information we will save as the file sizw could increase quickly, so probably a compression system will be necessary.