Save objects state to disk
An example of the file name would be simulation_12072018_124059.dat, the file contains a certain number of frames. Each frame has a header that tells how many objects there are, frame number, platform and game time. Each objects has a header that identifies it followed by the state of the object.